Add support for if/else gating behaviors
`glog` uses a type alias for `bool` to permit gating entire sections
of logic based on verbosity level. This is the syntax:

	if glog.V(2) {
		// ... do something
	}

Since `logif` is an interface to implementations, the API design has
to change slightly -- rather than aliasing bool and returning it for
the `V()` and `D()` functions, there are two additional functions
added to the interface, `IsV()` and `IsD()`. These functions return
`bool` permitting use of the same pattern, albeit with slightly less
elegance.